# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Constantine Mafia game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type PlayerCareerPath = typeof PlayerCareerPath[keyof typeof PlayerCareerPath];


export const PlayerCareerPath = {
  street_thug: 'street_thug',
  gangster: 'gangster',
  crime_boss: 'crime_boss',
  business_tycoon: 'business_tycoon',
} as const;

export interface Player {
  id: string;
  username: string;
  height: number;
  money: number;
  level: number;
  xp: number;
  careerPath: PlayerCareerPath;
  /** @nullable */
  currentMissionId: string | null;
  district: string;
  ownedAssetIds: string[];
  completedMissionIds: string[];
  createdAt: string;
  updatedAt: string;
}

export interface PlayerInput {
  username: string;
}

export interface PlayerLinkInput {
  username: string;
  height: number;
}

export type PlayerSaveCareerPath = typeof PlayerSaveCareerPath[keyof typeof PlayerSaveCareerPath];


export const PlayerSaveCareerPath = {
  street_thug: 'street_thug',
  gangster: 'gangster',
  crime_boss: 'crime_boss',
  business_tycoon: 'business_tycoon',
} as const;

export interface PlayerSave {
  username?: string;
  height?: number;
  money?: number;
  level?: number;
  xp?: number;
  careerPath?: PlayerSaveCareerPath;
  /** @nullable */
  currentMissionId?: string | null;
  district?: string;
  ownedAssetIds?: string[];
  completedMissionIds?: string[];
}

export type GameAssetCategory = typeof GameAssetCategory[keyof typeof GameAssetCategory];


export const GameAssetCategory = {
  vehicle: 'vehicle',
  property: 'property',
  weapon: 'weapon',
} as const;

export interface GameAsset {
  id: string;
  name: string;
  nameAr: string;
  nameFr: string;
  category: GameAssetCategory;
  subcategory: string;
  price: number;
  description: string;
  district: string;
  imageKey: string;
}

export interface PurchaseInput {
  assetId: string;
}

export type MissionDistrict = typeof MissionDistrict[keyof typeof MissionDistrict];


export const MissionDistrict = {
  ali_mendjeli: 'ali_mendjeli',
  centre_ville: 'centre_ville',
  old_city: 'old_city',
  ain_mlila: 'ain_mlila',
  airport: 'airport',
} as const;

export type MissionType = typeof MissionType[keyof typeof MissionType];


export const MissionType = {
  theft: 'theft',
  gang_trade: 'gang_trade',
  police_operation: 'police_operation',
  job: 'job',
  story: 'story',
} as const;

export type MissionDifficulty = typeof MissionDifficulty[keyof typeof MissionDifficulty];


export const MissionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  extreme: 'extreme',
} as const;

export interface Mission {
  id: string;
  title: string;
  titleAr: string;
  titleFr: string;
  description: string;
  descriptionAr: string;
  district: MissionDistrict;
  type: MissionType;
  difficulty: MissionDifficulty;
  reward: number;
  xpReward: number;
  requiredLevel: number;
  isStoryMission: boolean;
}

export interface MissionCompleteInput {
  playerId: string;
}

export interface MissionResult {
  success: boolean;
  moneyEarned: number;
  xpEarned: number;
  leveledUp: boolean;
  /** @nullable */
  newLevel: number | null;
  player: Player;
}

export interface LeaderboardEntry {
  rank: number;
  playerId: string;
  username: string;
  level: number;
  money: number;
  careerPath: string;
}

export type ListAssetsParams = {
category?: ListAssetsCategory;
};

export type ListAssetsCategory = typeof ListAssetsCategory[keyof typeof ListAssetsCategory];


export const ListAssetsCategory = {
  vehicle: 'vehicle',
  property: 'property',
  weapon: 'weapon',
} as const;

export type ListMissionsParams = {
district?: ListMissionsDistrict;
};

export type ListMissionsDistrict = typeof ListMissionsDistrict[keyof typeof ListMissionsDistrict];


export const ListMissionsDistrict = {
  ali_mendjeli: 'ali_mendjeli',
  centre_ville: 'centre_ville',
  old_city: 'old_city',
  ain_mlila: 'ain_mlila',
  airport: 'airport',
} as const;

export type GetLeaderboardParams = {
limit?: number;
};