# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Constantine Mafia game API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerCareerPath } from './playerCareerPath';

export interface Player {
  id: string;
  username: string;
  height: number;
  money: number;
  level: number;
  xp: number;
  careerPath: PlayerCareerPath;
  /** @nullable */
  currentMissionId: string | null;
  district: string;
  ownedAssetIds: string[];
  completedMissionIds: string[];
  createdAt: string;
  updatedAt: string;
}