# lib/api-zod/src/generated/types/playerLinkInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Constantine Mafia game API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerLinkInput {
  username: string;
  height: number;
}