# lib/api-zod/src/generated/types/playerSave.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Constantine Mafia game API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerSaveCareerPath } from './playerSaveCareerPath';

export interface PlayerSave {
  username?: string;
  height?: number;
  money?: number;
  level?: number;
  xp?: number;
  careerPath?: PlayerSaveCareerPath;
  /** @nullable */
  currentMissionId?: string | null;
  district?: string;
  ownedAssetIds?: string[];
  completedMissionIds?: string[];
}